support for cookie in alpha config and legacy file structure refactoring
