+++ pkg/apis/options/load_test.go
@@ -47,13 +47,26 @@ var _ = Describe("Load", func() {
 			InsecureOIDCSkipNonce: true,
 		},
 
+		LegacyCookie: LegacyCookie{
+			Name:           "_oauth2_proxy",
+			Secret:         "",
+			Domains:        nil,
+			Path:           "/",
+			Expire:         time.Duration(168) * time.Hour,
+			Refresh:        time.Duration(0),
+			Secure:         true,
+			HTTPOnly:       true,
+			SameSite:       "",
+			CSRFPerRequest: false,
+			CSRFExpire:     time.Duration(15) * time.Minute,
+		},
+
 		Options: Options{
 			ProxyPrefix:        "/oauth2",
 			PingPath:           "/ping",
 			ReadyPath:          "/ready",
 			RealClientIPHeader: "X-Real-IP",
 			ForceHTTPS:         false,
-			Cookie:             cookieDefaults(),
 			Session:            sessionOptionsDefaults(),
 			Templates:          templatesDefaults(),
 			SkipAuthPreflight:  false,
